+++ chess/src/main/java/pdp/model/ai/heuristics/BishopEndgameHeuristic.java
@@ -19,7 +19,7 @@ public class BishopEndgameHeuristic implements Heuristic {
    */
   @Override
   public float evaluate(final Board board, final boolean isWhite) {
-    int score = 0;
+    float score = 0;
     score += evaluateBishopMobility(board, true) - evaluateBishopMobility(board, false);
     score +=
         evaluateSameColorBishopsSamePlayer(board, true)
@@ -39,10 +39,10 @@ public class BishopEndgameHeuristic implements Heuristic {
    * @param isWhite true if white, false otherwise
    * @return score based on how many moves bishops can make
    */
-  private int evaluateBishopMobility(final Board board, final boolean isWhite) {
-    int score = 0;
+  private float evaluateBishopMobility(final Board board, final boolean isWhite) {
+    float score = 0;
     final BoardRepresentation bitboard = board.getBoardRep();
-    final int scoreForEachMove = 2;
+    final float scoreForEachMove = 2f;
     final List<Move> bishopMoves = bitboard.retrieveBishopMoves(isWhite);
 
     score += bishopMoves.size() * scoreForEachMove;
@@ -56,7 +56,7 @@ public class BishopEndgameHeuristic implements Heuristic {
    * @param isWhite true if white, false otherwise
    * @return Negative score if bishops are on the same color
    */
-  private int evaluateSameColorBishopsSamePlayer(final Board board, final boolean isWhite) {
+  private float evaluateSameColorBishopsSamePlayer(final Board board, final boolean isWhite) {
     final List<Position> bishops = board.getBoardRep().getBishops(isWhite);
     if (bishops.size() < 2) {
       return 0;
@@ -83,7 +83,7 @@ public class BishopEndgameHeuristic implements Heuristic {
    * @param isWhite true if white, false otherwise
    * @return Positive score if bishops are on the same color, 0 otherwise
    */
-  private int evaluateSameColorBishopsOpponent(final Board board, final boolean isWhite) {
+  private float evaluateSameColorBishopsOpponent(final Board board, final boolean isWhite) {
     final List<Position> bishopsPlayer1 = board.getBoardRep().getBishops(isWhite);
     final List<Position> bishopsPlayer2 = board.getBoardRep().getBishops(!isWhite);
     if (bishopsPlayer1.size() > 1 || bishopsPlayer2.size() > 1) {
@@ -111,8 +111,8 @@ public class BishopEndgameHeuristic implements Heuristic {
    * @param isWhite true if white, false otherwise
    * @return score for bishops that are close to the center
    */
-  private int evaluateCentralization(final Board board, final boolean isWhite) {
-    int score = 0;
+  private float evaluateCentralization(final Board board, final boolean isWhite) {
+    float score = 0;
     final List<Position> bishops = board.getBoardRep().getBishops(isWhite);
     final int noBonus = 0;
 
@@ -131,8 +131,8 @@ public class BishopEndgameHeuristic implements Heuristic {
    * @param isWhite true if white, false otherwise
    * @return Negative score if bishop is blocked by pawns
    */
-  private int evaluateBadBishop(final Board board, final boolean isWhite) {
-    int score = 0;
+  private float evaluateBadBishop(final Board board, final boolean isWhite) {
+    float score = 0;
     final List<Position> bishops = board.getBoardRep().getBishops(isWhite);
     final List<Position> pawns = board.getBoardRep().getPawns(isWhite);
